Rename fetch_types.d.ts to dom_types.d.ts
