+++ BUILD.gn
@@ -59,9 +59,9 @@ ts_sources = [
   "js/console.ts",
   "js/deno.ts",
   "js/dispatch.ts",
+  "js/dom_types.d.ts",
   "js/errors.ts",
   "js/fetch.ts",
-  "js/fetch_types.d.ts",
   "js/global-eval.ts",
   "js/globals.ts",
   "js/libdeno.ts",
