+++ js/dom_types.d.ts
